requirements.txt
pandas
sqlalchemy
psycopg2-binary
mysql-connector-python
openpyxl
